Include the cost of store population.
Previously we populated the store and then ran an experiment.
But this feels unnecessarily unfair to the join measurement.
(The push system does work in the population phase that the
join system does later.) So include the population phase. This
is not ideal either; what would be ideal needs more thought.