working versions of join annotations for allowing staleness and turning off automatic maintenance.
by default all joins are automatically maintined for the ranges
that have been validated. changes to the input ranges
are expanded to the output range and the store is upated.

Join::set_maintained(false) turns off automatic maintenance.
this means that the join will be fully validated for the requested
range on each validation query. in terms of the implementation,
it means that there is no validjoin installed in sink_ranges_ and
no copy ranges installed in source_ranges_.

Join::set_staleness(double s) sets the expiration time of the
join expansion to now + s. this means that future validation
requests that fall within this staleness bound that overlap
the expanded range will use the existing expansion and will
not re-validate the stale range. it also implies that there
is no automatic maintenance of the expanded range. in terms
of the implementation, there is a validjoin installed into
sink_ranges_, but it is given an expiration time. no
copy ranges are installed into source_ranges_. when a new
validation request is processed, the expiration of the
stale validjoin is checked and it is either used as is or
re-validated.

i have some thoughts on what these annotations do to the
invariants of the store_ (w.r.t. the validity of individual keys)
but i will save that for email.