Correct total_size
Reminder: A ServerRangeSet contains up to four classes of range.
1. Fully active ranges: i.e., range's beginning point is at or before
   the set's current position, and range's ending point is at or after
   the set interval's ending point.
2. Currently-inactive ranges: i.e., range's beginning point is after
   the set's current position.
3. Currently-active, future-inactive ranges: i.e., range's end point
   is within the set's interval, but after the current position.
4. Dead ranges: i.e., range's ending point is at or before the the
   set's current position.

sw_ has a 1 bit in the positions corresponding to #2 and #3. The
total_size() method should return the total number of intervals.
We incorrectly assumed that, if sw_!=0, then there must be intervals
in set #2. But that's not necessarily right; and intervals in set