add memcached client
Memcached takes 47 seconds in user mode, but returns 140 times more
tweets due to lack of get offset.

$./obj/pqserver -d --push
5000 USERS HAVE SUBSCRIBERS:
  zero: 0  mean: 109.553  sdev: 188.125
    min 2, 1% 5, 10% 8, 25% 11, 50% 17, 75% 105, 90% 397, 99% 790, max 899
{"nposts":45059,"nfull":14863,"nupdate":965078,"nposts_read":761069188,"time":47.394962}

The implementation of PQTwitter now includes the overhead of marshalling
and unmarshalling, which is necessary if we have a RPC client. Besides,
the cost of marshalling and unmarshalling is only 4% (see below). So why not.

$./obj/pqserver
5000 USERS HAVE SUBSCRIBERS:
  zero: 0  mean: 109.553  sdev: 188.125
  min 2, 1% 5, 10% 8, 25% 11, 50% 17, 75% 105, 90% 397, 99% 790, max 899
{
        "nposts":45059,
        "nfull":14863,
        "nupdate":965078,
        "nposts_read":5046241,
        "time":13.424839,
        "store_size":5450219,
        "source_ranges_size":5000,
        "sink_ranges_size":5001
}