adding parameters to operators (source ranges and accumulators).
as an example. the operator jvt_count_match now takes parameters 'lbound' and 'ubound' params that will setup a filter on the keys that will be counted.
the value of the key to be counted is first interpreted as an integer an compared to the given bounds. if it matches, it is counted. oherwise
it is discarded. it also works if a key that was previously counted is erased or if a key is modified, causing it to go to/from being counted.

i plan to implement something similar for the jvt_copy_match operator.

this is not the best way to customize operators. it is limited in functionality (what if i want to filter based on a value regex match?) but it shows promise.