let each SourceRange remembers the sink range for each resultkey_
This improves test_karma_online by 10%, but in an unexpected way.
I was hoping that store_.insert_check could use sb.hint() if sb.hint()
is the targeted iterator. But no! Will check it out later.

It doesn't improve twitter - twitter is slowed down by 7%, probably
due to the overhead of updating sinkrange. The reason
is that twitter couldn't benefit from per-SourceRange sinkrange. Rather,
twitter requires a SinkBound at a higher level - a SinkBound for
"t|<user>" for each <user>.