Flush pull (non-maintained) data from the store.
Previously scanning a non-maintained join would leave data in the
store potentially forever. Now we record the non-maintained join range
as a "flushable" range, and then flush it, i.e., remove data from the
store, on the very next validate.

This may be relatively slow but that may not matter.