Replace an expansion key with an expansion "context".
A context is a compressed match: a set of full slot assignments.
Contexts are used instead of keys in the SourceRange expansion
process. This is because they are smaller and because they apply to
intermediate ranges too, which key assignments don't.

This commit does not improve semantic behavior. It just adds a ton of
interfaces. :|