using builtin pull support
builtin pull doubles the runtime of pull mode comparing with the
original implementation of pull in pqrwmicro.cc. For the following
micro-twitter,

$ ./obj/pqserver --nusers=2000 --nops=2000000 --rwmicro --nfollower=20
--prefresh=5 --no-push

the runtime of previous pull implementation is 19 seconds.
Now, it is 35 seconds. The major reason is that the server doesn't
flush keys in t table in time, which slows the subsequent pull request
(of a different scan range) down. With this partially 'fixed' (i.e. let
SinkRange::flush erase all keys in the table), the runtime is
22 seconds. A complete 'fix' might be to provide a Table::flush API
so that pqrwmicro could flush the table after each timeline refresh,
and might reduce the runtime to 19 seconds.