Filters: An alternate query execution strategy.
A filter is used only in pull mode (but it could easily work in push
mode too). It is applied AFTER we reach the sink, not before. Are we
becoming a database? Sure! This makes --celebrity3=850 1sec faster,
because if no celebrity tweets, then we don't ever look at the cs|
table.