use db params.
add flush calls. this doesn't really work because it is a single flush of the dbpool buffer.
more queries could be added to the buffer after the flush (by a tamed method that was
waiting for the response from a previous query to finish). we can then hang waiting for
events to be fired that are stuck in the dbpool buffer. we need an external check (like a timer)
or some kind of flushing loop?