Add an optional bound to the Table::iterator so that Table::iterator::fix() will not go outside the bounds of a given subtable.
This is a fix for a bug that was manifesting when triecuts were used
and a sink range was validated in a cold cache. The iterators
returned when validating the sink and source ranges were calling
fix(), which would walk the entire table structure (which contained
a whole lot of empty subtables) until it reached the supertable's
end. This was very expensive when there were millions of empty
subtables.

We ideally want to stay within the bounds of table_for(first, last)
in these cases, so that is what this fix accomplishes. It allows
the iterators returned by Table::lower_bound to only iterate
over the subtable structure bounded by some root table.

We can still iterate over the entire store, but if a scan range
is confined to a single subtable, as it often is when using triecuts,
the iterator will be bounded appropriately.

Developers should use some care to compare iterators that are derived from
the same subtable. The codebase has been updated so that there is consistent
usage of Server::table_for() when assigning iterators. In addition,
the method Table::iterator::table_end() will return an end iterator that is
appropriately bounded as to avoid infinite loops when used as a sentinel.

This fixes the bug with respect to the cold cache,
which caused the test program to run for hours. Unit tests work,
and twitter tests work when run locally and remotely.