Fix a bug in sink validtion that manifested as an infinite loop.
A sink could be invalidated while it was being validated! This
could happen if SinkRange::validate_step calls Table::validate on an
input range and that range was previously evicted. As part of
the cleanup of the evicted range, its dependents are invalidated,
possibly including the sink that is being re-validated
(because it was still in the results vector of the source range
and is marked as valid because it is being re-validated).

The solution was to only invalidate a sink if it is not currently
being validated. The source range believes the sink is valid
and calls Sink::invaldiate as part of SourceRange::invaldiate.
Thus, inside Sink::invalidate we only take action if we are
not in the middle of a re-valdiation.

Another options is to only mark the sink as valid after it completes
a validation attempt, but a bunch of stuff already checks Sink::valid
and we want to keep that in place. This is a smaller change.